// submodules/common-dto/constants/imageTarget.enum.ts
enum ImageTargetEnum {
    // User
    UserAvatar = 'user-avatar',
    BackgroundImage = 'background-image',
    Document = 'document',
    // Artwork
    Artwork = 'artwork',
    // Misc
    Keyword = 'keyword',
    Category = 'category',
    Material = 'material',
    Style = 'style',
    Subject = 'subject',
    // Locations
    CountryFlag = 'country-flag',
}

export default ImageTargetEnum;